+++ src/data/bedrock/block/convert/BlockSerializerDeserializerRegistrar.php
@@ -117,15 +117,17 @@ use pocketmine\data\bedrock\block\convert\BlockStateSerializerHelper as Helper;
 use pocketmine\data\bedrock\block\convert\BlockStateWriter as Writer;
 use pocketmine\data\bedrock\block\convert\property\BoolFromStringProperty;
 use pocketmine\data\bedrock\block\convert\property\BoolProperty;
+use pocketmine\data\bedrock\block\convert\property\CommonProperties;
 use pocketmine\data\bedrock\block\convert\property\DummyProperty;
+use pocketmine\data\bedrock\block\convert\property\EnumFromRawStateMap;
 use pocketmine\data\bedrock\block\convert\property\FlattenedCaveVinesVariant;
-use pocketmine\data\bedrock\block\convert\property\FlattenedIdModel;
+use pocketmine\data\bedrock\block\convert\property\IntFromRawStateMap;
 use pocketmine\data\bedrock\block\convert\property\IntProperty;
-use pocketmine\data\bedrock\block\convert\property\Model;
 use pocketmine\data\bedrock\block\convert\property\OptionSetFromIntProperty;
 use pocketmine\data\bedrock\block\convert\property\StringProperty;
 use pocketmine\data\bedrock\block\convert\property\ValueFromIntProperty;
 use pocketmine\data\bedrock\block\convert\property\ValueFromStringProperty;
+use pocketmine\data\bedrock\block\convert\property\ValueMappings;
 use pocketmine\math\Facing;
 use function array_map;
 use function count;
