+++ src/data/bedrock/block/convert/property/IntFromRawStateMap.php
@@ -21,7 +21,7 @@
 
 declare(strict_types=1);
 
-namespace pocketmine\data\bedrock\block\convert;
+namespace pocketmine\data\bedrock\block\convert\property;
 
 use function array_flip;
 use function is_array;
