+++ src/data/bedrock/block/convert/property/StateMap.php
@@ -21,7 +21,7 @@
 
 declare(strict_types=1);
 
-namespace pocketmine\data\bedrock\block\convert;
+namespace pocketmine\data\bedrock\block\convert\property;
 
 /**
  * @phpstan-template TValue
