+++ src/data/bedrock/block/convert/property/ValueFromStringProperty.php
@@ -26,7 +26,6 @@ namespace pocketmine\data\bedrock\block\convert\property;
 use pocketmine\data\bedrock\block\BlockStateDeserializeException;
 use pocketmine\data\bedrock\block\convert\BlockStateReader;
 use pocketmine\data\bedrock\block\convert\BlockStateWriter;
-use pocketmine\data\bedrock\block\convert\StateMap;
 use function array_keys;
 use function array_map;
 use function strval;
